+++ src/main/java/entier/person/sale/dto/req/BienTheReq.java
@@ -8,6 +8,6 @@ public class BienTheReq {
     String hinhAnh;
     String mauSac;
     String kichCo;
-    Double gia;
+    Long gia;
     Integer tonKho;
 }
